First version of integrating React Virtuoso to display large tables.
