+++ db/python/layers/billing.py
@@ -1,16 +1,14 @@
-from typing import Any
-
 from db.python.layers.bq_base import BqBaseLayer
 from db.python.tables.bq.billing_ar_batch import BillingArBatchTable
 from db.python.tables.bq.billing_daily import BillingDailyTable
 from db.python.tables.bq.billing_daily_extended import BillingDailyExtendedTable
 from db.python.tables.bq.billing_gcp_daily import BillingGcpDailyTable
 from db.python.tables.bq.billing_raw import BillingRawTable
-from models.enums import BillingSource, BillingTimeColumn, BillingTimePeriods
+from models.enums import BillingSource
 from models.models import (
+    BillingBatchCostRecord,
     BillingColumn,
     BillingCostBudgetRecord,
-    BillingHailBatchCostRecord,
     BillingTotalCostQueryModel,
 )
 
@@ -206,7 +204,7 @@ class BillingLayer(BqBaseLayer):
     async def get_cost_by_ar_guid(
         self,
         ar_guid: str | None = None,
-    ) -> list[dict]:  # BillingHailBatchCostRecord:
+    ) -> list[BillingBatchCostRecord]:  # BillingHailBatchCostRecord:
         """
         Get Costs by AR GUID
         """
@@ -220,49 +218,18 @@ class BillingLayer(BqBaseLayer):
         ) = await ar_batch_lookup_table.get_batches_by_ar_guid(ar_guid)
 
         if not batches:
-            return BillingHailBatchCostRecord(
-                ar_guid=ar_guid,
-                batch_ids=[],
-                costs=[],
-            )
+            return []
 
-        # Then get the costs for the given AR GUID/batches from the main table
-        all_cols = [BillingColumn.str_to_enum(v) for v in BillingColumn.raw_cols()]
-
-        all_cols.append(BillingColumn.AR_GUID.value)
-
-        if ar_guid is None:
-            filters = {
-                'batch_id': batches,
-            }
-        else:
-            filters = {
-                'ar_guid': ar_guid,
-            }
-
-        query = BillingTotalCostQueryModel(
-            fields=all_cols,
-            source=BillingSource.EXTENDED,
-            start_date=start_day.strftime('%Y-%m-%d'),
-            end_date=end_day.strftime('%Y-%m-%d'),
-            filters=filters,
-            group_by=False,
-            time_column=BillingTimeColumn.DAY,
-            time_periods=BillingTimePeriods.DAY,
+        billing_table = self.table_factory(BillingSource.EXTENDED)
+        results = await billing_table.get_batch_cost_details(
+            start_day, end_day, batches, ar_guid
         )
-
-        billing_table = self.table_factory(query.source, query.fields)
-        records = await billing_table.get_total_cost(query)
-        return BillingHailBatchCostRecord(
-            ar_guid=ar_guid,
-            batch_ids=batches,
-            costs=records,
-        )
+        return results
 
     async def get_cost_by_batch_id(
         self,
         batch_id: str | None = None,
-    ) -> list[Any]:  # BillingHailBatchCostRecord:
+    ) -> list[BillingBatchCostRecord]:  # BillingHailBatchCostRecord:
         """
         Get Costs by Batch ID
         """
@@ -278,43 +245,12 @@ class BillingLayer(BqBaseLayer):
             batches,
         ) = await ar_batch_lookup_table.get_batches_by_ar_guid(ar_guid)
 
-        # if not batches:
-        #     return BillingHailBatchCostRecord(ar_guid=ar_guid, batch_ids=[], costs=[])
+        if not batches:
+            return (
+                []
+            )  # BillingHailBatchCostRecord(ar_guid=ar_guid, batch_ids=[], costs=[])
 
-        # # Then get the costs for the given AR GUID/batches from the main table
-        # all_cols = [BillingColumn.str_to_enum(v) for v in BillingColumn.raw_cols()]
-
-        # all_cols.append(BillingColumn.AR_GUID.value)
-        # all_cols.append(BillingColumn.BATCH_ID.value)
-
-        # if ar_guid is None:
-        #     filters = {
-        #         'batch_id': batches,
-        #     }
-        # else:
-        #     filters = {
-        #         'ar_guid': ar_guid,
-        #     }
-
-        # query = BillingTotalCostQueryModel(
-        #     fields=all_cols,
-        #     source=BillingSource.EXTENDED,
-        #     start_date=start_day.strftime('%Y-%m-%d'),
-        #     end_date=end_day.strftime('%Y-%m-%d'),
-        #     filters=filters,
-        #     group_by=False,
-        #     time_column=BillingTimeColumn.DAY,
-        #     time_periods=BillingTimePeriods.DAY,
-        # )
-        # billing_table = self.table_factory(query.source, query.fields)
-        # records = await billing_table.get_total_cost(query)
-        # return BillingHailBatchCostRecord(
-        #     ar_guid=ar_guid,
-        #     batch_ids=batches,
-        #     costs=records,
-        # )
         billing_table = self.table_factory(BillingSource.EXTENDED)
-
         results = await billing_table.get_batch_cost_details(
             start_day, end_day, batches, ar_guid
         )
