+++ web/package.json
@@ -30,6 +30,7 @@
         "react-responsive": "^9.0.2",
         "react-router-dom": "^6.0.1",
         "react-syntax-highlighter": "^15.4.4",
+        "react-virtuoso": "^4.6.3",
         "recharts": "^2.8.0",
         "remark-gfm": "^3.0.1",
         "remark-toc": "^8.0.1",
